Breaking change: Rename root folder to render-markdown, changes require statement
## Details

The original name markdown has a larger potential for naming collisions,
to avoid this have decided to fix this problem early on and rename the
root folder which changes the require statement to require('render-markdown')
instead of require('markdown'). Sorry to any early users :(

The plugin name as a whole markdown.nvim can also collide, have
updated instructions for lazy to specify a name, though this is not
necessary in all cases.