Use the current buffer parser and recurse all trees rather than parsing buffer as markdown
## Details

This should solve issue: https://github.com/MeanderingProgrammer/markdown.nvim/issues/3

This will allow for markdown to be rendered with this plugin when
injected in other languages, for example Python comments.

Currently we attempt to parse the entire buffer as markdown which
leads to some strange behavior, i.e. interpretting single line
comments as markdown.

After this change we will read the injected markdown sections.

This should also allow us to take advantage of the markdown_inline
parser for later changes, which is nice.