Improve support for LaTeX multi-line formulas
## Details

As a follow up from:
https://github.com/MeanderingProgrammer/markdown.nvim/issues/6,
when LaTeX formulas are multi-line we end up showing the formula
on the first line and not overwriting the actual contents, which
does not look great.

Rather than attempting to overlay the formula as before, use
virtual lines above the text itself.

This is different from the standard rendering experience of an overlay
but I think is generally a nicer experience.

This will have some issues if the inline LaTeX is surrounded by other
text, which seems like an okay trade off.