Add support for dynamically changing conceallevel between rendered and raw views
# Details

Brought up in https://github.com/MeanderingProgrammer/markdown.nvim/issues/10,
it would be cool if we could conceal links when rendering markdown.

Turns out if conceallevel is set to a high enough value this happens by default.

Hook into this existing feature be increasing conceallevel to 3 when
rendering markdown, and use user default value in all other cases.

Allow individuals to opt out of this by changing the conceallevel
values for both default and rendered.