Split adding marks / virtual text into individual handlers by language. Add cache for latex expressions.
# Details

This change primarily functions as an internal refactor.

Changes the UI module to mostly be a top level orchestrator and
moves main rendering logic into language specific handlers.

This makes changing indivual parts easier without intermingling
any logic only needed for a specific part.

This made it easier to cache the encodings from pylatexenc to
avoid encoding the same expression multiple times without adding
a cache to the entire UI module which is only needed for the latex
component.