Allow for different width headings
# Details

As stated in this issue: https://github.com/MeanderingProgrammer/markdown.nvim/issues/12,
we currently only handle headings with a string width of 1 or 2, longer
headings will overlay the titles.

To fix this calculate the width of the heading to determine the amount
of paddings needed. This will still fail if the heading provided is
longer than the level of the heading, since we have fewer characters
to work with.