Do not render bullet points before checkboxes
# Details

Resolves: https://github.com/MeanderingProgrammer/markdown.nvim/issues/18

When the next sibling of a list marker is a bullet point overlay
with spaces rather than spaces + bullet point.

We could also do a conceal here to have the text shift over rather
than adding spaces, however this has the downside of a conceal where
once the cursor enters the line the text jumps back over and the marker
becomes visible again. I prefer to avoid the text jumping behavior.