Use strdisplaywidth in place of # for string lengths
Resolves: https://github.com/MeanderingProgrammer/markdown.nvim/issues/26

Currently there are a couple places where when calculating the
length of a particular string we simply use the Lua `#`. This works
well enough for most characters, but not well with more complex ones
such as with Chinese, as described in the issue.

Using strdisplaywidth accounts for how Neovim will actually render the
characters.