Add support for user defined handlers
# Details

Allow users to specify handlers for whatever language they want
using the `custom_handlers` configuration parameter.

As an implementation it is meant to give users the same ability
as the native renderers for `markdown`, `markdown_inline` and
`latex`. Convenience methods are not provided and users are
expected to use the vim api directly.

Provide an example as well as some details in the README.

By integrating with this plugin the `extmarks` set by the user
will behave identically to the ones created by the plugin. All
validation on buffers, windows and state also works as expected.