Support highlighting callout quote markers
resolves: https://github.com/MeanderingProgrammer/markdown.nvim/issues/24

# Details

Figuring out which callout, if any, is in a `block_quote` is non trivial
using treesitter as the links in a `block_quote` are markdown_inline while
the markers themselves are standard markdown.

Passing information between these 2 parsers is not really possible, i.e.
we cannot take a markdown `TSNode` and query for markdown_inline elements.

Re-parsing the contents of the node is likely possible though I imagine there
would need to be some handling of (start/end) (row/col).

To get around this compexity we take an approach that's really similar to
how we compute list levels to pick a bullet point.

When handling a `quote_marker` find the closest parent `block_quote` element.
From there we get the text of the `block_quote` node and do a naive
string contains check to determine if the `block_quote` is a callout and,
if so, which kind it is. Based on this select an appropriate highlight.

This likely has edge cases just based on how adhoc the logic is, though,
at the moment, I cannot think of any. Release this into the wild and
hopefully any bugs will be reported so I can iron out the kinks.