Add latex parser as optional and update health check
## Details

The requirements to get latex support working for this plugin are
not clear as shown by: https://github.com/MeanderingProgrammer/markdown.nvim/issues/32.

Update the README to add the missing need for the latex parser.

More importantly the current health check logic only checked for
the minimum needs of this plugin as there was no way to indicate
whether the user wanted latex support or not.

Add a `latex_enabled` flag which defaults to true. Functionally it
does allow enabling / disabling the latex handler but it is mainly
used in the updated health check logic to check for the additional
requirements needed for latex, i.e. the parser and the latex2text
executable. Setting the flag to false allows users to remove these
warnings if they do not want latex support by this plugin.