Get mode at time of event instead of at time of render
resolves: https://github.com/MeanderingProgrammer/markdown.nvim/issues/36

# Details

Since all rendering and clearing of `extmarks` happens through a call to
`vim.schedule` it is possible that a mode change occurred between a
triggering event and the response to that event.

In most cases this is not really a problem, however some plugins will
programmatically cause several mode changes in quick succession which
can cause this plugin to be out of sync, i.e. thinking that a render is
not necessary since the last mode was already rendered, however when
rendering the last mode, the mode had changed and a render did not occur.

An example of this is `fzf-lua` which uses `terminal` mode to perform
certain actions.

This can cause the plugin to fall behind the current user state, however
rendering should be fast enough for this not to be a problem. Can revisit
this approach if it proves to be a problem.