Add ability for custom handlers to extend builtin
# Details

Currently custom handlers can only replace the logic of builtin
handlers. Meaning adding one feature to the markdown handler is
not possible without replacing all of the existing logic.

Allow users to specify a new field in each handler called extends,
which defaults to false. When set to true both the user defined
handler as well as the builtin one will run.

This can lead to contradictions in behavior, this is not handled
in any way and I don't really think there is a good way to do this.
Ideally these custom handlers offer functionality that does not overlap
with any of the existing functionality.

Make some other minor changes as well such as where the disable latex
logic lives.

Update demos and update script to handle setting the handler interface
in the README.