Feature: add language icon above code blocks
# Details

Idea from: https://github.com/OXY2DEV/markview.nvim

Adds support for overlaying a language icon above code blocks
using nvim-web-devicons.

Make the nvim-web-devicons dependency optional and gracefully
handle the case when it is not installed using a pcall.

Add a code_style option to allow users to disable this if they
prefer not to have it.