Use full mode when deciding to render or not
resolves: https://github.com/MeanderingProgrammer/markdown.nvim/issues/43

# Details

Currently we truncate mode to the first character when deciding to
render or not. This reduces the amount of control we have, as described
in the issue above, where not rendering during operator pending mode
`no` would improve navigation.

The fix is to remove any truncation logic so we access the full mode.

This on its own handles operator pending correctly as `no` mode is
no longer considered a render mode. So the defaults do not change.

There could be some edge cases that crop up with different mode
combinations so may need to update the default set of rendered modes,
but seems to work as expected from some basic use.