Fix language selector to pull language from info_string
resolves: https://github.com/MeanderingProgrammer/markdown.nvim/issues/44

# Details

Previously when getting the langauge for a fenced code block we simply
pulled the info_string in full. However `markdown` supports additional
information in this component so would not resolve correctly. Instead
we need to pull the language node from the info_string node to handle
these cases.

To do this add a separate query / capture to handle getting the language
of a fenced code block.

Since the language can now be in different places, including nested in
curly braces generating the right virtual text becomes harder. To avoid
dealing with these cases use an inline extmark to shift things for us.
This does mean that the icon feature is now restricted to neovim
versions >= 0.10.0. Add a version check to avoid failing for early
versions.