Attach most auto commands at buffer level
## Details

This functions as a performance improvement by removing most of the
global auto commands setup by this plugin. Primarily common events
like TextChanged / ModeChanged.

Instead of having these at a global level and having internal logic
to avoid running on certain buffers keep only the FileType auto command
at a global level, use the input file types as a pattern directly.

From there attach the other auto commands to the buffer scope. Add
a manager module to keep track of buffers we've attached to. This is
to enable performing global actions on only buffers we care about. In
this case for WinResized events and for the toggle command.

Users should see no visible change from this other than improved
background performance of this plugin.