Add support for code blocks in callouts / quotes
# Details

The current logic for finding quote markers in a block quote relies
on several tree-sitter queries that do not handle different element
types well, an example of this is code blocks.

To fix this rather than having a top level query which pulls quote
marker nodes use 2 separate queries. This first gets the block_quote,
the second runs on the block_quote node and gets all continuations,
regardless of how they are nested.

This does to a certain extent allow arbitrary nesting, however the
highlights chosen for nested callouts do not look great, but it does
work, so that's cool.

Other minor changes:
- README order
- Add checkhealth field to bug report template
- Store callouts as a list so order is consistent