Refactor! Likely to break all configs, group properties by component
## Details

This is a massive breaking change that modifies nearly every
configuration parameter.

Give the amount of changes if you would rather punt dealing with this
for now feel free to use the 'v3.3.1' git tag which will keep you at
the previous commit.

For example with lazy.nvim:

```lua
{
    'MeanderingProgrammer/markdown.nvim',
    tag = 'v3.3.1',
    ...
}
```

Most configs will likely still work, in that the plugin will not
break, but it will essentially ignore all user settings. There are
known settings that if set previously will break the plugin:

- `dash`
- `checkbox.unchecked`
- `checkbox.checked`
- `quote`

The idea of this change is rather than scattering properties for
say 'headings' through the top level 'headings' parameter and a
separate 'highlights.headings' parameter, we instead create a top
level 'heading' parameter that stores icons and highlights. This
should make the problem of what to modify easier as well as allow
individual components to evolve more easily.

List of all changes and how to fix them split by section below.

### Base

- `start_enabled` -> `enabled`

### Latex

- `latex_enabled` -> `latex.enabled`
- `latex_converter` -> `latex.converter`
- `highlights.latex` -> `latex.highlight`

### Headings

- `headings` -> `heading.icons`
- `highlights.heading.backgrounds` -> `heading.backgrounds`
- `highlights.heading.foregrounds` -> `heading.foregrounds`

### Code

- `code_style` -> `code.style`
- `highlights.code` -> `code.highlight`

### Dash

- `dash` -> `dash.icon`
- `highlights.dash` -> `dash.highlight`

### Bullets

- `bullets` -> `bullet.icons`
- `highlights.bullet` -> `bullet.highlight`

### Checkbox

- `checkbox.unchecked` -> `checkbox.unchecked.icon`
- `highlights.checkbox.unchecked` -> `checkbox.unchecked.highlight`
- `checkbox.checked` -> `checkbox.checked.icon`
- `highlights.checkbox.checked` -> `checkbox.checked.highlight`

### Quote

- `quote` -> `quote.icon`
- `highlights.quote` -> `quote.highlight`

### Table

- `table_style` -> `pipe_table.style`
- `cell_style` -> `pipe_table.cell`
- `highlight.table.head` -> `pipe_table.head`
- `highlight.table.row` -> `pipe_table.row`

### Callouts

- `callout.note` -> `callout.note.rendered`
- `callout.tip` -> `callout.tip.rendered`
- `callout.important` -> `callout.important.rendered`
- `callout.warning` -> `callout.warning.rendered`
- `callout.caution` -> `callout.caution.rendered`
- `highlights.callout.note` -> `callout.note.highlight`
- `highlights.callout.tip` -> `callout.tip.highlight`
- `highlights.callout.important` -> `callout.important.highlight`
- `highlights.callout.warning` -> `callout.warning.highlight`
- `highlights.callout.caution` -> `callout.caution.highlight`
- `callout.custom.*` -> `callout.*` (i.e. unnest from custom block)

### Others

Any remaing changes are covered within that component.

I.e. `code_style` is covered in Code, `highlights.table` is covered in
Table, `highlights.callout.note` is covered in Callouts, etc.