Iterate through table children instead of using query
## Details

Currently we use separate queries for processing different parts
of a markdown pipe table.

Instead, since we already get the top level table, we can iterate
through the children which will give us the exact same nodes as
from the captures.

Update the default query to remove the unneeded capture groups.

Add a NodeInfo class to store text value and range in table.