feature: Improve table & code conceal behavior
## Details

Use the current conceal level + metadata to determine how to
render language headers and table cells.

For code blocks we simply check if the langauge element is likely
concealed and if it is not only render an icon isntead of an
icon + the language name.

For table cells things are a little more complicated, but boils
down to using `inline` extmarks to pad the table cells such that
they take up as much space as if nothing were concealed. This
is done through a new table cell style 'padded' which is also the
new default.

Since this relies on inline extmarks it only works with neovim
>= 0.10.0, but will not break for earlier versions, it just will
not perform the padding operation.

Since cells can be any inline code the extmarks we add which shift
text also need to be accounted for in addition to the conceals. This
is easy enough for now since we do not use inline marks much, only
for links really. But it is a point of fragility and means that
table rendering needs to know about the functionality of everything.
Still this is much better than the obtrusive overlay, which users
can continue to use if they prefer.