fix: Raw tables match full when rendered uniformally
## Details

The current logic for rendering a full table only checks the
raw text to make sure the widths line up.

This makes sense for 'overlay' & 'padded' tables but not so
much for 'raw' tables.

Rather than using the same computed width use the width after
text has been concealed and inline elements have been added.

This is essentially the inverse operation of padded. Where for
padded tables we fill any removed space, for raw tables we
use the removed space as the width.

Other minor changes:

- Add a class to each module to improve LSP diagnostics, immediately
  calls out invalid method calls
- Move width calculation into str module, will make changing it in
  the future easier as well as making consistency more likely
- Refactor table rendering into multiple top level methods
- Add a shared module for resolving link icons in the same way for
  markdown & markdown_inline handlers