feature: Use SignColumn background for signs allow them to be disabled by buftype
# Details

Request: https://github.com/MeanderingProgrammer/markdown.nvim/issues/58

Not all color schemes use the Normal background color for the sign
column. This can lead to some not so great looking icons. Fix this
by combining the color we were going to set with the SignColumn
background. Add a new configuration section 'sign' where this is
defined and moifiable.

While we are adding a new section for signs add 2 other abilities.

- The ability to globally disable them from components that create them
- The ability to disable them within a set of buftypes, use a default
  value of 'nofile' so that signs do not appear in LSP documentation