feature: Add border style to code blocks, use thin by default
## Details

Request: https://github.com/MeanderingProgrammer/markdown.nvim/issues/62

Rather than using a highlight group for the entire range of code blocks
allow half height characters to be used when start and end of code
block are entirely concealed.

This involves using the inverse of the code highlight group and
repeating the configured characters across the width.

Allow using the original implementation with `{ border = 'thick' }`.

When rendering the body we now need to be aware of how the language
was rendered, since placing a language above the code block interferes
with the half height row. To fix this we need to handle the language
from the top level code block logic, instead of using 2 separate
capture groups for code block & language.

Add sign configuration to README as well.