feat: Add dash.width option
## Details

Add dash.width option which is either the string 'full' or an integer.

When it is 'full' the dashed line will extend across the entire line,
which is the current behavior and the default value.

When it is a number the width of the dashed line will be the hard coded
value provided.

Original PR: https://github.com/MeanderingProgrammer/markdown.nvim/pull/92

Co-Authored-By: Zeioth <Zeioth@hotmail.com>