fix: Marks on current row should only be hidden for current buffer
## Details

Currently when LSP signature help is opened the first line virtual
text is hidden. This is because the cursor for that buffer is on
the first line, however this should not matter since that buffer is
not active.

To fix this check whether the buffer being rendered is the current
buffer when getting the cursor position. If it is not the current
buffer use nil as the row value and render marks when row is nil.

Refactor the logic for showing vs. hiding a mark into a local function
rather than a large if statement. This lets us comment what's going
on and I have a feeling that logic will need to evolve in other ways
over time.

Add a todo doc to keep track of various ideas that I haven't fully
ruled out implementing.

Add plugin configuration and checkboxes of confirmations as
fields to bug report template.