feat: Add heading.width option
## Details

Currently headings always extend to the end of the current line.

Some users do not like that, to support this add a width attribute
with 2 values, 'full' & 'block'.

'full' behaves the same as now and extends the background to the end
of the line. 'block' on the other hand ends right after the text of
the heading.

Original PR: https://github.com/MeanderingProgrammer/markdown.nvim/pull/94

Co-Authored-By: Zeioth <Zeioth@hotmail.com>