feat: Improve debug logging
## Details

After all the recent changes the current logging approach does not
capture all the necessary information at the right time.

To get all the necessary information in addition to logging nodes we'll
need to also log the marks that we produce. Doing this around the same
time as node will make it easier to understand what produced what. There
is already a function we call as we encounter new marks our own list.add.
Rename this method to add_mark and add a logging statement, now we will
be logging marks as they are generated right around the same time as
the node producing the marks, sweet.

To run at the right time we need to use the fact that most refresh cycles
are due to cursor movements so all that would get logged are option value
changes. To fix this change where the logger starts and ends to only
occur on a parse refresh. Remove logging from setting options since
this now happens outside of this cycle.

Other minor changes include:

- Add a name attribute to all log entries and change the format slightly
- Always making the name of parameters optional instead of sometimes doing
  it to the type and other times the name