feat: Add alignment indicator icons to tables
## Details

Add configurable alignment_indicator field to pipe_table. The
indicator is positioned in the left, right, or center of every
column with alignment information. The indicator must be exactly
a single display width.

While the change is simple at a high level, to implement it we
needed to really parse through the delimiter row. Previously we
did some very simply gsub replacement on the entire line. But now
we need to compute the width between pipes to accurately measure
what's underneath since cells themselves are free to occupy as
much or as little space as they like.

As a result a parser needed to be added which mostly iterates through
the delimiter row and does a good amount of sanity checking before
doing anything with the table.

This information can likely be used to improve how tables are handled
in the future, for now we just add the indicator. Alignmnet itself
is completely up to the user.