feature: auto-setup using plugin directory
## Details

Request: https://github.com/MeanderingProgrammer/markdown.nvim/issues/79

This allows users to use this plugin without calling the setup function.

In this case the default configuration will be used.

In the original request the suggestion was to use ftplugin, however since
the filetypes this plugin runs on are configurable this is not possible,
hence the usage of the plugin directory instead.

This change should have no impact to existing user while adding
functionality to avoid the setup paradigm if desired. As a consequence
of other changes that needed to be made the concepts of configuration
and initialization are much better separated instead of being coupled.

In order to support this some minor structural changes needed to be made:

- Color and Manager modules as well as User Command are initialed inside
  plugin directory instead of setup to avoid creating multiple copies.
- FileType autocommand can no longer use a pattern directly since it
  is not known until after user calls setup. Instead the callback now
  runs for every event and we filter out events related to other filetypes.
- Creating user command in plugin directory it makes sense to move
  available commands into a separate API module. Use a metatable to allow
  users to call the API methods without changing require logic.
- Minimal init for unit tests now needs to be ordered and call runtime
  method for this plugin. This mimics what a plugin manager with
  dependencies would do.

Other minor changes:

- Add limitation of not running in Telescope Preview
- Add todo to remove gifs from being inside the repo
- Some unrelated comments added to keep track of things