fix: Remove gifs from repo & fix conceal on window change
## Details

I didn't really consider the fact that storing gif demos in this
repo would make the repo and especially its history much larger
and forces everyone to clone mostly useless data.

Add the gifs to gitignore. For now will manually upload them to
GitHub when they change and update the README. Less convenient
but not too bad.

Currently if you swith out of a markdown buffer the line the
cursor was last on will remain anti-concealed. Add BufLeave
event which will recognize that the buffer is no longer active
and show all marks.

Split buffer autocommands into 2 groups, one that requires re-parsing,
currently only TextChanged. And another that does not, which is now
BufWinEnter and BufLeave. This works because the ui will still parse
marks if it does not have any cached.

Add cache invalidation to the setup method. Since we use the plugin
directory many users are likely calling it twice, once from plugin
and again when configuring. To avoid showing old marks which may not
match the configuration the invalidation logic was added. This is an
unlikely edge case and was really only necessary for some of the unit
tests to continue functioning.