feat(dev): add profiler to get render time data
## Details

Add profiler module that wraps the main `refresh` method where
all rendering logic lives.

This only tracks info if the `profile` field has been set to true,
which should only be the case for myself but I guess others can
if they would like, could help debug performance issues in the
future.

Tracks runtime in milliseconds of the various different kind of
possible refreshes, i.e. movement vs full parse at a buffer level.

Add an api called stats that processes the runtime data and outputs
some basic info (min, max, mean) to `stats.json` in whatever directory
I'm currently in.

May add a benchmark suite to use this in the future, currently it
is to help me get a baseline of where we're at.

Other unrelated change to types to define classes directly on tables
rather than defining a class and adding it to a table.