feat!: Support full buftype options
## Details

Adds support for an `overrides` field that allows for essentially a full
configuration per buftype. If no configuration is specified for a
particular buftype or a partial configuration the logic falls back to
the top level configuration. The top level configuration can be viewed
as the default and then overrides change it for specific buffers.

Because this mechanism is much more powerful it deprecates all other
buftype options as they can now be accomplished via this new mechanism.

The equivalents are as follows:

- `exclude.buftypes.<value>` -> `overrides.buftype.<value>.enabled = false`
- `sign.exclude.buftypes.<value>` -> `overrides.buftype.<value>.sign.enabled = false`

The previous default behavior of not showing signs on LSP floating
docs has been ported over, so anyone using default configurations
should not experience any problems.

Since the configuration is accessed all through the codebase there is a
pretty large code change here, but a lot if it is passing the new
resolved configuration around.

This enables a feature like different code padding by default vs. in LSP
floating windows with:

```lua
require('render-markdown').setup({
    code = { left_pad = 4 },
    overrides = {
        buftype = {
            nofile = {
                code = { left_pad = 0 },
            },
        },
    },
})
```