fix: Store conceal nodes based on row to improve overlap calculation
## Details

The last change introduced a regression in the event that a buffer
has many instances of concealed nodes, for example many inline
code blocks or links.

The fix for this is straightforward, rather than storing nodes as
a flat array split the array based on row. Makes the lookup much
faster since most nodes get filtered out.