feat(perf): Improve hiding marks under cursor
## Details

Currently on every cursor movement we clear all buffer marks
then add all the marks back one at a time avoiding the ones
at the current cursor position.

Instead what we can do is avoid touching marks that are unaffected,
i.e. all of the ones that remain rendered, render any previously
removed marks, and remove the relevant marks for the current line.

To accomplish this have added an Extmark class that wraps the raw
mark information from our parsers. These marks store the id of the
extmark from the call to nvim_buf_set_extmark. Now when we attempt
to render a mark we check if we should, i.e. is it on the current
row or not.

- If we should render the mark only do it if it does not have an
  id set and then we set the id
- If we should not render the mark only do it if it does have an
  id set and then we unset the id

By doing this we still loop through all the marks but we only touch
marks that change from shown -> hidden or from hidden -> shown.

Other minor changes:

- Rename `request` module to `context`
- Add benchmark for medium table