fix: Setting custom configuration for vim-plug
## Details

Bug report: https://github.com/MeanderingProgrammer/markdown.nvim/issues/111

After the transition to initialization through the plugin directory we
essentially made the plugin configuration dependent on initialization
order. Whichever thing happens to call the setup function last
determines the configuration.

This works for plugin managers like lazy.nvim since the plugin directory
is run before the user calls setup, so user settings will override the default.

However this is not the case for vim-plug which runs setup before the plugin
directory. Side not, I believe this is the way neovim is meant to work and
lazy.nvim / other managers change this ordering.

To fix this I have modified the main setup function to pick up the last
non empty configuration.

That way if plugin first runs like with lazy.nvim any later calls to setup
by the user will override the configuration, there is no need to re-initialize
if an empty configuration is supplied.

Now if setup runs first it'll initialize the state and since the plugin
directory provides an empty configuration it will be skipped instead of
overwriting the user supplied configuration.