chore!: Complete deprecation of custom handlers render method
## Details

Completes the removal of `render` method in `custom_handlers`.

Originally slated for 2024-08-19, I'll be out around this time for a few
days and want to get this plugin into a steady state after the rename.

Get the lua rock updated figure out what to do with the old one, etc.

I did double check the feature is unused, at least from what's available on Github:

- https://github.com/search?q=custom_handlers+render+++language%3ALua&type=code
- https://github.com/search?q=custom_handlers+markdown++language%3ALua&type=code
- https://github.com/search?q=custom_handlers+MeanderingProgrammer++language%3ALua&type=code

No one appears to be using it which gives me enough confidence to remove it early